Complete Refactor (#26)
* WIP: moving route handlers to new files

* Refactoring the rest of the route handlers and models

* Fixing formatting

* Finishing up compliance with PEP8
