--- aswwu/__init__.py
@@ -1,2 +1,0 @@
-# ignore this file
-# it's a necessary *empty* file to tell python to treat this folder as a package
