+++ aswwu/alchemy.py
@@ -1,15 +1,22 @@
 # alchemy.py
 
-from sqlalchemy.orm import sessionmaker,joinedload
-from sqlalchemy import create_engine
-
 # import and set up the logging
 import logging
+
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker, joinedload
+import aswwu.models.bases as base
+import aswwu.models.mask as mask_model
+from aswwu.archive_models import ArchiveBase
+
+Base = base.Base
+ElectionBase = base.ElectionBase
+PagesBase = base.PagesBase
+JobsBase = base.JobsBase
+
 logger = logging.getLogger("aswwu")
 
 # import the necessary models (all of them in this case)
-from aswwu.models import *
-from aswwu.archive_models import *
 
 # defines the databases URLs relative to "server.py"
 engine = create_engine("sqlite:///../databases/people.db")
@@ -27,44 +34,44 @@ JobsBase.metadata.create_all(jobs_engine)
 # bind instances of the databases to corresponding variables
 Base.metadata.bind = engine
 dbs = sessionmaker(bind=engine)
-s = dbs()
+people_db = dbs()
 # same for archives
 ArchiveBase.metadata.bind = archive_engine
 archive_dbs = sessionmaker(bind=archive_engine)
-archive_s = archive_dbs()
-#same for elections
+archive_db = archive_dbs()
+# same for elections
 ElectionBase.metadata.bind = election_engine
 election_dbs = sessionmaker(bind=election_engine)
-election_s = election_dbs()
-#same for pages
+election_db = election_dbs()
+# same for pages
 PagesBase.metadata.bind = election_engine
 pages_dbs = sessionmaker(bind=pages_engine)
-page_s = pages_dbs()
+page_db = pages_dbs()
 
 JobsBase.metadata.bind = election_engine
 jobs_dbs = sessionmaker(bind=jobs_engine)
-jobs_s = jobs_dbs()
+jobs_db = jobs_dbs()
 
 
 # updates a model, or creates it if it doesn't exist
-def addOrUpdate(thing):
+def add_or_update(thing):
     try:
-        s.add(thing)
-        s.commit()
+        people_db.add(thing)
+        people_db.commit()
         return thing
     except Exception as e:
         logger.info(e)
-        s.rollback()
+        people_db.rollback()
 
 
 # finds all rows for a given model
 def query_all(model):
     thing = None
     try:
-        thing = s.query(model).all()
+        thing = people_db.query(model).all()
     except Exception as e:
         logger.info(e)
-        s.rollback()
+        people_db.rollback()
     return thing
 
 
@@ -72,21 +79,21 @@ def query_all(model):
 def query_by_wwuid(model, wwuid):
     thing = None
     try:
-        thing = s.query(model).filter_by(wwuid=str(wwuid)).all()
+        thing = people_db.query(model).filter_by(wwuid=str(wwuid)).all()
     except Exception as e:
         logger.info(e)
-        s.rollback()
+        people_db.rollback()
     return thing
 
 
 # finds all rows for a given model matching the given ID
-def query_by_id(model, id):
+def query_by_id(model, aid):
     thing = None
     try:
-        thing = s.query(model).filter_by(id=id).first()
+        thing = people_db.query(model).filter_by(id=aid).first()
     except Exception as e:
         logger.info(e)
-        s.rollback()
+        people_db.rollback()
     return thing
 
 
@@ -94,16 +101,16 @@ def query_by_id(model, id):
 def query_by_field(model, field, value):
     thing = None
     try:
-        thing = s.query(model).filter(getattr(model, field).like(value)).all()
+        thing = people_db.query(model).filter(getattr(model, field).like(value)).all()
     except Exception as e:
         logger.info(e)
-        s.rollback()
+        people_db.rollback()
     return thing
 
 
 # finds a user with the given WWUID
 def query_user(wwuid):
-    thing = query_by_wwuid(User, str(wwuid))
+    thing = query_by_wwuid(mask_model.User, str(wwuid))
     if thing:
         thing = thing[0]
     return thing
@@ -112,129 +119,111 @@ def query_user(wwuid):
 # permanently deletes a given model
 def delete_thing(thing):
     try:
-        s.delete(thing)
-        s.commit()
+        people_db.delete(thing)
+        people_db.commit()
     except Exception as e:
         logger.info(e)
-        s.rollback()
+        people_db.rollback()
 
 
-def query_all_Election(model):
+def query_all_election(model):
     thing = None
     try:
-        thing = election_s.query(model).all()
+        thing = election_db.query(model).all()
     except Exception as e:
         logger.info(e)
-        election_s.rollback()
+        election_db.rollback()
     return thing
 
 
-def addOrUpdateElection(thing):
+def add_or_update_election(thing):
     try:
-        election_s.add(thing)
-        election_s.commit()
+        election_db.add(thing)
+        election_db.commit()
         return thing
     except Exception as e:
         logger.info(e)
-        election_s.rollback()
+        election_db.rollback()
 
 
 # finds all rows for a given model matching the given WWUID
-def query_by_wwuid_Election(model, wwuid):
+def query_by_wwuid_election(model, wwuid):
     thing = None
     try:
-        thing = election_s.query(model).filter_by(wwuid=str(wwuid)).all()
+        thing = election_db.query(model).filter_by(wwuid=str(wwuid)).all()
     except Exception as e:
         logger.info(e)
-        election_s.rollback()
+        election_db.rollback()
     return thing
 
 
 # updates a model, or creates it if it doesn't exist
-def addOrUpdatePage(thing):
+def add_or_update_page(thing):
     try:
-        page_s.add(thing)
-        page_s.commit()
+        page_db.add(thing)
+        page_db.commit()
         return thing
     except Exception as e:
         logger.info(e)
-        page_s.rollback()
+        page_db.rollback()
 
 
 def query_by_page_url(model, url):
     thing = None
     try:
-        thing = page_s.query(model).options(joinedload('*')).filter_by(url=str(url)).all()
+        thing = page_db.query(model).options(joinedload('*')).filter_by(url=str(url)).all()
     except Exception as e:
         logger.info(e)
-        page_s.rollback()
+        page_db.rollback()
     return thing
 
 
 def query_by_page_id(model, page_id):
     thing = None
     try:
-        thing = page_s.query(model).options(joinedload('*')).filter_by(id=str(page_id)).all()
+        thing = page_db.query(model).options(joinedload('*')).filter_by(id=str(page_id)).all()
     except Exception as e:
         logger.info(e)
-        page_s.rollback()
+        page_db.rollback()
     return thing
-
-
-def query_by_page_url(model, url):
-    thing = None
-    try:
-        thing = page_s.query(model).options(joinedload('*')).filter_by(url=str(url)).all()
-    except Exception as e:
-        logger.info(e)
-        page_s.rollback()
-    return thing
-
-# def query_page(page_id):
-#     thing = None
-#     try:
-#         thing = query_by_page_id(User, str(page_id))
-#         if thing:
-#             thing = thing[0]
-#     except Exception as e:
-#         logger.info(e)
-#         page_s.rollback()
-#     return thing
 
 
 # updates a model, or creates it if it doesn't exist
-def addOrUpdateForm(thing):
+def add_or_update_form(thing):
     try:
-        jobs_s.add(thing)
-        jobs_s.commit()
+        jobs_db.add(thing)
+        jobs_db.commit()
         return thing
     except Exception as e:
         logger.info(e)
-        jobs_s.rollback()
+        jobs_db.rollback()
+
 
 def query_by_job_name(model, name):
     thing = None
     try:
-        thing = jobs_s.query(model).options(joinedload('*')).filter_by(job_name=str(name)).all()
+        thing = jobs_db.query(model).options(joinedload('*')).filter_by(job_name=str(name)).all()
     except Exception as e:
         logger.info(e)
-        jobs_s.rollback()
+        jobs_db.rollback()
     return thing
 
-def query_all_Forms(model):
+
+def query_all_forms(model):
     thing = None
     try:
-        thing = jobs_s.query(model).all()
+        thing = jobs_db.query(model).all()
     except Exception as e:
         logger.info(e)
-        jobs_s.rollback()
+        jobs_db.rollback()
     return thing
 
+
 # permanently deletes a given model
-def delete_thing_Forms(thing):
+def delete_thing_forms(thing):
     try:
-        jobs_s.delete(thing)
-        jobs_s.commit()
+        jobs_db.delete(thing)
+        jobs_db.commit()
     except Exception as e:
         logger.info(e)
-        jobs_s.rollback()
+        jobs_db.rollback()
