+++ aswwu/base_handlers.py
@@ -1,35 +1,44 @@
 # base_handlers.py
 
-import tornado.web
-import logging
-import requests
-import json
-import time
 import datetime
-import hmac
-import base64
 import hashlib
+import hmac
+import json
+import logging
+import time
+
+import requests
+import tornado.web
+
 from settings import testing
 
-# import modles and alchemy functions as needed
-from aswwu.models import *
-from aswwu.alchemy import *
+# import models and alchemy functions as needed
+import aswwu.models.mask as mask_model
+import aswwu.alchemy as alchemy
 
 logger = logging.getLogger("aswwu")
 
+
 # model used only in this file
 # acts as an extension of the User and Profile models
+def import_profile(profile, exported_json):
+    for field in exported_json:
+        if exported_json[field]:
+            setattr(profile, field, exported_json[field])
+
+
 class LoggedInUser:
     def __init__(self, wwuid):
         self.wwuid = wwuid
-        profile = query_by_wwuid(Profile, wwuid)
-        user = query_user(wwuid)
+        profile = alchemy.query_by_wwuid(mask_model.Profile, wwuid)
+        user = alchemy.query_user(wwuid)
         if len(profile) == 0:
-            old_profile = archive_s.query(globals()['Archive' + get_last_year()]).filter_by(wwuid=str(wwuid)).all()
-            new_profile = Profile(wwuid=str(wwuid), username=user.username, full_name=user.full_name)
+            old_profile = alchemy.archive_db.query(globals()['Archive' + get_last_year()]).\
+                filter_by(wwuid=str(wwuid)).all()
+            new_profile = mask_model.Profile(wwuid=str(wwuid), username=user.username, full_name=user.full_name)
             if len(old_profile) == 1:
-                self.import_profile(new_profile, old_profile[0].export_info())
-            profile = addOrUpdate(new_profile)
+                import_profile(new_profile, old_profile[0].export_info())
+            profile = alchemy.add_or_update(new_profile)
         else:
             profile = profile[0]
         self.username = user.username
@@ -42,12 +51,8 @@ class LoggedInUser:
         self.status = user.status
 
     def to_json(self):
-        return {'wwuid': str(self.wwuid), 'username': str(self.username), 'full_name': str(self.full_name), 'photo': self.photo, 'roles': str(','.join(self.roles)), 'status': str(self.status)}
-
-    def import_profile(self, profile, exported_json):
-        for field in exported_json:
-            if exported_json[field]:
-                setattr(profile, field, exported_json[field])
+        return {'wwuid': str(self.wwuid), 'username': str(self.username), 'full_name': str(self.full_name),
+                'photo': self.photo, 'roles': str(','.join(self.roles)), 'status': str(self.status)}
 
 
 # this is the root/base handler for all other handlers
@@ -55,7 +60,8 @@ class BaseHandler(tornado.web.RequestHandler):
     # newer JS frameworks send an OPTIONS request as a "preflight" to check if the server is safe
     # this just tells the framework that it is safe to send data here
     def options(self, *path_args, **path_kwargs):
-        self.set_header("Access-Control-Allow-Headers", "Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With, token")
+        self.set_header("Access-Control-Allow-Headers",
+                        "Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With, token")
         pass
 
     # allow data to come from anywhere
@@ -63,24 +69,24 @@ class BaseHandler(tornado.web.RequestHandler):
         self.set_header("Access-Control-Allow-Origin", "*")
 
     # creates an HMAC digest that is a hexadecimal hash based on a provided message
-    def generateHMACDigest(self, message):
+    def generate_hmac_digest(self, message):
         secret = self.application.settings['secret_key']
         signature = hmac.new(secret, message, digestmod=hashlib.sha256).hexdigest()
         return signature
 
     # create a authorization token for the given WWUID based on the current time
-    def generateToken(self, wwuid):
+    def generate_token(self, wwuid):
         now = int(time.mktime(datetime.datetime.now().timetuple()))
         message = str(wwuid)+"|"+str(now)
-        return message+"|"+self.generateHMACDigest(message)
+        return message+"|"+self.generate_hmac_digest(message)
 
     # see if the authorization token received from the user has been tampered with (i.e. copied or stolen)
-    def validateToken(self, token):
+    def validate_token(self, token):
         token = token.split("|")
         if len(token) != 3:
             return False
-        compareTo = self.generateHMACDigest(token[0]+"|"+token[1])
-        return compareTo == token[2]
+        compare_to = self.generate_hmac_digest(token[0] + "|" + token[1])
+        return compare_to == token[2]
 
     # global hook that allows the @tornado.web.authenticated decorator to function
     # checks for an authorization header and attempts to validate the user with that information
@@ -94,14 +100,14 @@ class BaseHandler(tornado.web.RequestHandler):
                 else:
                     token = self.get_cookie("token")
                     wwuid = token.split("|")[0]
-                    dateCreated = int(token.split("|")[1])
+                    date_created = int(token.split("|")[1])
                     now = int(time.mktime(datetime.datetime.now().timetuple()))
                     # check if token was created with the last 2 weeks (14 days) and is a valid token
-                    if (now - dateCreated) > (60 * 60 * 24 * 14) or not self.validateToken(token):
+                    if (now - date_created) > (60 * 60 * 24 * 14) or not self.validate_token(token):
                         user = None
                     else:
                         user = LoggedInUser(wwuid)
-            except Exception as e:
+            except:
                 user = None
 
             return user
@@ -111,13 +117,16 @@ class BaseHandler(tornado.web.RequestHandler):
     def prepare(self):
         # some modern JS frameworks force data to be sent as JSON
         # this isn't a bad thing at all, just requires us to "prepare" the data before it's used internally
-        if "Content-Type" in self.request.headers and self.request.headers["Content-Type"].startswith("application/json") and len(self.request.arguments) == 0:
+        if "Content-Type" in self.request.headers and self.request.headers["Content-Type"].\
+                startswith("application/json") and len(self.request.arguments) == 0:
             try:
                 json_data = json.loads(str(self.request.body))
-                for key in json_data: json_data[key] = [json_data[key]]
+                for key in json_data:
+                    json_data[key] = [json_data[key]]
                 self.request.arguments = json_data
-            except Exception as e:
+            except:
                 pass
+
 
 # effectively useless, but at least provides an endpoint for people accessing "/" by accident
 class BaseIndexHandler(BaseHandler):
@@ -126,23 +135,25 @@ class BaseIndexHandler(BaseHandler):
         user = self.current_user
         self.write(user.to_json())
 
+
 # login and/or register users as needed
 class BaseLoginHandler(BaseHandler):
     # verify login information with the WWU servers somehow
-    def loginWithWWU(self, username, password):
+    @staticmethod
+    def login_with_wwu(username, password):
         # NOTE: this url is old, yet still functional
         # expect it to change eventually
         mask_url = "https://www.wallawalla.edu/auth/mask.php"
 
         # pass the username and password to our accepted URL
-        r = requests.post(mask_url, data = {'username': username, 'password': password}, verify=True)
+        r = requests.post(mask_url, data={'username': username, 'password': password}, verify=True)
 
         # parse out the ugly response
-        parsedUser = r.text.encode('utf-8')[6:-7]
-        parsedUser = json.loads(parsedUser)['user']
-        if parsedUser:
-            parsedUser['wwuid'] = parsedUser['wwcid']
-            return parsedUser
+        parsed_user = r.text.encode('utf-8')[6:-7]
+        parsed_user = json.loads(parsed_user)['user']
+        if parsed_user:
+            parsed_user['wwuid'] = parsed_user['wwcid']
+            return parsed_user
         else:
             return None
 
@@ -155,7 +166,10 @@ class BaseLoginHandler(BaseHandler):
     # the main login/registration handler
     def post(self):
         logger.debug("'class':'LoginHandler','method':'post', 'message': 'invoked'")
-        self.write({'error':'We\'ve switched over to the university login. Try refreshing the page and clearing your cache to login with the new method. If you\'re having more issues email aswwu.webmaster@wallawalla.edu'})
+        self.write({'error': 'We\'ve switched over to the university login.'
+                             ' Try refreshing the page and clearing your cache to login with the new method.'
+                             ' If you\'re having more issues email aswwu.webmaster@wallawalla.edu'})
+
 
 # verify a user's authorization token
 class BaseVerifyLoginHandler(BaseHandler):
@@ -166,7 +180,7 @@ class BaseVerifyLoginHandler(BaseHandler):
         user = self.current_user
         if user:
             # if a user exists, refresh their token for them
-            token = self.generateToken(user.wwuid)
+            token = self.generate_token(user.wwuid)
             self.write({'user': user.to_json(), 'token': token})
             self.set_cookie("token", token, domain='.aswwu.com', expires_days=14)
         else:
@@ -177,47 +191,3 @@ class BaseVerifyLoginHandler(BaseHandler):
 def get_last_year():
     year = tornado.options.options.current_year
     return str(int(year[:2]) - 1) + str(int(year[2:4]) - 1)
-
-#
-# # login and/or register users as needed
-# class SAMLLoginHandler(BaseHandler):
-#     # if someone gets here they have bigger problems than not being logged in
-#     def get(self):
-#         logger.debug("not logged in")
-#         self.write({'error': 'not logged in'})
-#
-#     # the main login/registration handler
-#     def post(self):
-#         logger.debug("'class':'LoginHandler','method':'post', 'message': 'invoked'")
-#         username = self.get_argument('username', None)
-#
-#         # make sure we actually received something from the user
-#         if username and password:
-#             try:
-#                 # expects a dictionary to be returned here (JSON)
-#                 user_dict = self.loginWithWWU(username, password)
-#                 if user_dict:
-#                     # lookup the user
-#                     user = query_user(user_dict['wwuid'])
-#                     if not user:
-#                         # if a matching user doesn't exist, create it
-#                         user = User(wwuid=user_dict['wwuid'], username=user_dict['username'], full_name=user_dict['fullname'], status=user_dict['status'])
-#                         addOrUpdate(user)
-#                     # generate a new token for this login
-#                     token = self.generateToken(user_dict['wwuid'])
-#                     # create a new LoggedInUser model
-#                     user = LoggedInUser(user_dict['wwuid'])
-#                     # this worked out, send it all back to the user
-#                     self.write({'user': user.to_json(), 'token': str(token)})
-#                 else:
-#                     # self.loginWithWWU didn't return what we expected
-#                     logger.info("LoginHandler: error "+r.text)
-#                     self.write({'error':'invalid login credentials'})
-#             except Exception as e:
-#                 # you've got some debugging to get through if you're here
-#                 logger.error("LoginHandler exception: "+ str(e.message))
-#                 self.write({'error': str(e.message)})
-#         else:
-#             # tell the user to send some better information
-#             logger.error("LoginHandler: invalid post parameters")
-#             self.write({'error':'invalid post parameters'})
